Fix numeric literal formatting
Prior to this commit, in some situations, formatting changes the datatype of numeric literals (xsd:double and xsd:decimal). The change happens if the lexical form is ambiguous because it does not include a fractional part.

This commit checks for this condition and uses the quoted form in these cases.